Small fixes

+++ scoop/rogue/views/actions/flag.py
@@ -1,7 +1,12 @@
 # coding: utf-8
 from django.contrib import messages
+from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import user_passes_test
 from django.utils.translation import ugettext_lazy as _
+
+from scoop.content.models.content import Content
+from scoop.core.templatetags.text_tags import humanize_join
+from scoop.core.util.data.typeutil import make_iterable
 from scoop.core.util.signals import record
 from scoop.messaging.models.alert import Alert
 from scoop.messaging.models.thread import Thread
@@ -9,38 +14,48 @@ from scoop.user.util.auth import is_staff
 
 
 @user_passes_test(is_staff)
-def ban_and_alert(request, user, template):
-    """ Bannir le profil et alterter les utilisateurs """
-    action_done = user.profile.ban()
-    # Créer la notice à l'utilisateur
-    if action_done is True:
-        recipients = Thread.objects.related_users(user)
-        Alert.objects.alert(recipients, 'user.profile.fake', 'security', {'profile': user.profile})
-        count = len(recipients)
-        notice = _("%(user)s has been banned and reported to %(count)d users.") % {'user': user, 'count': count}
-        messages.success(request, notice)
-        record.send(None, request.user, 'user.ban-alert.user', user)
+def ban_and_alert(request, users):
+    """ Bannir les profils et alterter les utilisateurs qui ont été en contact """
+    users, banned = make_iterable(users), []
+    if users:
+        for user in users:
+            if user.profile.ban():
+                recipients = Thread.objects.related_users(user, ack=True)
+                Alert.objects.alert(recipients, 'user.profile.fake', 'security', {'profile': user.profile})
+                record.send(get_user_model(), request.user, 'user.ban-alert.user', user)
+                banned.append(user)
+        if banned:
+            messages.success(request, humanize_join(banned, 3, _("user has been banned;users hae been banned")))
+            return True
+    return False
 
 
 @user_passes_test(is_staff)
-def remove_pictures(request, user):
+def remove_pictures(request, users):
     """ Retirer toutes les images d'un profil """
-    pictures = user.profile.get_pictures()
-    count = pictures.count()
-    if count > 0:
-        Alert.objects.alert(user, 'user.profile.picture.invalid', 'notification', {'count': count, 'pictures': pictures})
-        notice = _("%(count)s pictures by %(user)s have been deleted.") % {'count': count, 'user': user}
-        messages.success(request, notice)
-    else:
-        messages.warning(request, _("No picture was deleted."))
-    pictures.delete()
-    record.send(None, request.user, 'content.delete.picture', user)
+    users, destroyed = make_iterable(users), []
+    if users:
+        for user in users:
+            pictures = user.profile.get_pictures()
+            count = pictures.count()
+            if count > 0:
+                Alert.objects.alert(user, 'user.profile.picture.invalid', 'notification', {'count': count, 'pictures': pictures})
+                record.send(get_user_model(), request.user, 'content.delete.picture', user)
+                pictures.delete()
+                destroyed.append(user)
+        if destroyed:
+            messages.success(request, humanize_join(destroyed, 3, _("user has images removed;users have images removed")))
+            return True
+    messages.info(request, _("No picture has to be deleted."))
+    return False
 
 
 @user_passes_test(is_staff)
 def lock_content(request, content):
     """ Verrouiller un contenu contre les modifications """
-    content.lock()
-    Alert.objects.alert(list(content.authors.all()), 'content.content.lock', 'notification', {'content': content})
-    messages.success(request, _("The content was protected from user interaction."))
-    record.send(None, request.user, 'content.lock.content', content)
+    if content.lock():
+        Alert.objects.alert(content.get_authors(), 'content.content.lock', 'notification', {'content': content})
+        messages.success(request, _("The content was locked and protected from user interaction."))
+        record.send(Content, request.user, 'content.lock.content', content)
+        return True
+    return False
